refactor: refactor stage-3
Use volume for PostgreSQL in the docker-compose.yaml. Add the health
check parameters to the mta.yaml. Allow unauthenticated access to the
/actuator/health endpoint. Update the project version. Add static imports.
Add the basic authentication credentials to the application.yaml. Update
the definitions of filter chains inside the security configuration classes.
Update the usage of testcontainers and enable security filters in the tests.